Preliminary FTP support
Right now data I/O is possible with the encryption scheme mostly
respected, except that in this build FTP is used (instead of FTPS). To
be fixed in future commit(s). Also, a byte "0D" is randomly added, which
makes I/O fail randomly as well. To be fixed, too.